Add files via upload
I'm sorry for this being so late. There is a time on there but its very basic and its in integer for which I thought would be easy but if its necessary for it to be in timer float format ill change it. Tonight ill be working on the health bars.